requirements.txt
bz2file==0.98
pandas==2.0.3
Pillow==10.1.0
requests==2.31.0
streamlit==1.26.0
